+++ react/__tests__/index.test.tsx
@@ -19,6 +19,7 @@ import {
   viewCartWithItemsMock,
   viewCartWithNoItemsMock,
 } from '../__mocks__/viewCart'
+import { transaction } from '../__mocks__/transaction'
 
 jest.mock('../modules/utils/shouldSendGA4Events')
 
@@ -415,61 +416,10 @@ describe('GA4 events', () => {
 
   describe('purchase', () => {
     it('sends an event that signifies a successful checkout on orderPlaced page', () => {
-      const transactionProducts = [
-        {
-          attachments: [],
-          brand: 'New Offers!!',
-          brandId: '2000045',
-          category: 'Apparel & Accessories',
-          categoryId: '25',
-          categoryIdTree: ['25'],
-          categoryTree: ['Apparel & Accessories'],
-          components: [],
-          ean: '9812983',
-          id: '9',
-          measurementUnit: 'un',
-          name: 'Top Everyday Necessaire 100 RMS',
-          originalPrice: 1600.99,
-          price: 1600.99,
-          priceTags: [],
-          productRefId: '',
-          quantity: 1,
-          seller: 'VTEX',
-          sellerId: '1',
-          sellingPrice: 1600.99,
-          sku: '20',
-          skuName: '100 RMS',
-          skuRefId: '9812983',
-          slug: 'everyday-necessaire',
-          tax: 0,
-          unitMultiplier: 1,
-        },
-      ]
+      const data = transaction
 
-      const data = {
-        accountName: 'storecomponents',
-        corporateName: null,
-        currency: 'USD',
-        event: 'orderPlaced',
-        eventName: 'vtex:orderPlaced',
-        openTextField: null,
-        orderGroup: '1310750551387',
-        ordersInOrderGroup: ['1310750551387-01'],
-        salesChannel: '1',
-        transactionAffiliation: 'VTEX',
-        transactionCurrency: 'USD',
-        transactionCustomTaxes: {},
-        transactionDate: '2023-02-14T18:56:47.0019167Z',
-        transactionDiscounts: 0,
-        transactionId: '1310750551387',
-        transactionLatestShippingEstimate: '2023-02-15T18:56:52.1493235Z',
-        transactionPayment: { id: 'FCE420A6B22C45D3BD60FD5DB55D34D1' },
-        transactionShipping: 1942.61,
-        transactionProducts,
-        transactionSubtotal: 1600.99,
-        transactionTax: 0,
-        transactionTotal: 3543.6,
-      }
+      data.event = 'orderPlaced'
+      data.eventName = 'vtex:orderPlaced'
 
       const message = new MessageEvent('message', { data })
 
@@ -661,6 +611,42 @@ describe('GA4 events', () => {
       })
     })
   })
+
+  describe('refund', () => {
+    it('sends an event when the user refunds an order', () => {
+      const data = transaction
+
+      data.event = 'refund'
+      data.eventName = 'vtex:refund'
+
+      const message = new MessageEvent('message', { data })
+
+      handleEvents(message)
+
+      expect(mockedUpdate).toHaveBeenCalledWith('refund', {
+        ecommerce: {
+          coupon: null,
+          currency: 'USD',
+          items: [
+            {
+              item_brand: 'New Offers!!',
+              item_category: 'Apparel & Accessories',
+              item_id: '9',
+              item_name: 'Top Everyday Necessaire',
+              item_variant: '20',
+              price: 1600.99,
+              quantity: 1,
+            },
+          ],
+          shipping: 1942.61,
+          tax: 0,
+          transaction_id: '1310750551387',
+          value: 3543.6,
+        },
+      })
+    })
+  })
+
   describe('add_shipping_info', () => {
     it('sends an event when a user add a shipping info', () => {
       const data = shippingInfoMock
@@ -685,6 +671,93 @@ describe('GA4 events', () => {
               price: 197.99,
             },
           ],
+        },
+      })
+    })
+  })
+
+  describe('add_to_wishlist', () => {
+    it('sends an event when the user add a product to wishlist', () => {
+      const message = new MessageEvent('message', { data: productDetails })
+
+      handleEvents(message)
+
+      expect(mockedUpdate).toHaveBeenCalledWith('view_item', {
+        ecommerce: {
+          currency: 'USD',
+          value: 1540.99,
+          items: [
+            {
+              item_id: '16',
+              item_name: 'Classic Shoes Top',
+              item_list_name: 'List of products',
+              item_brand: 'Mizuno',
+              item_variant: '35',
+              price: 1540.99,
+              quantity: 2000000,
+              discount: 0,
+              item_category: 'Apparel & Accessories',
+              item_category2: 'Shoes',
+            },
+          ],
+        },
+      })
+    })
+  })
+  describe('login', () => {
+    it('sends an event when a user login on store with any method', () => {
+      const data = {
+        event: 'login',
+        eventName: 'vtex:login',
+        method: 'Google',
+      }
+
+      const message = new MessageEvent('message', { data })
+
+      handleEvents(message)
+
+      expect(mockedUpdate).toHaveBeenCalledWith('login', {
+        ecommerce: {
+          method: 'Google',
+        },
+      })
+    })
+  })
+  describe('search', () => {
+    it('sends an event when a user search for a term on store', () => {
+      const data = {
+        event: 'search',
+        eventName: 'vtex:search',
+        term: 'Top Wood',
+      }
+
+      const message = new MessageEvent('message', { data })
+
+      handleEvents(message)
+
+      expect(mockedUpdate).toHaveBeenCalledWith('search', {
+        ecommerce: {
+          search_term: 'Top Wood',
+        },
+      })
+    })
+  })
+
+  describe('sign_up', () => {
+    it('sends an event when a user signup on store with any method', () => {
+      const data = {
+        event: 'signUp',
+        eventName: 'vtex:signUp',
+        method: 'Google',
+      }
+
+      const message = new MessageEvent('message', { data })
+
+      handleEvents(message)
+
+      expect(mockedUpdate).toHaveBeenCalledWith('sign_up', {
+        ecommerce: {
+          method: 'Google',
         },
       })
     })