+++ CHANGELOG.md
@@ -23,6 +23,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for sending GA4 [`begin_checkout`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#begin_checkout) when `vtex:beginCheckout` is received.
 - Support for sending GA4 [`view_cart`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#view_cart) when `vtex:beginCheckout` is received.
 - Support for sending GA4 [`add_shipping_info`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#add_shipping_info) when `vtex:addShippingInfo` is received.
+- Support for sending GA4 [`search`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#search) when `vtex:beginCheckout` is received.
+- Support for sending GA4 [`login`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#login) when `vtex:login` is received.
+- Support for sending GA4 [`refund`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#refund) when `vtex:refund` is received.
+- Support for sending GA4 [`add_to_wishlist`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#add_to_wishlist) when `vtex:addToWishlist` is received.
+- Support for sending GA4 [`sign_up`](https://developers.google.com/analytics/devguides/collection/ga4/reference/events#sign_up) when `vtex:signUp` is received.
 
 ## [3.4.0] - 2023-02-15